Define components on their route
